Changes to comments and contents in markdown cells
Nice work on the notebooks! This was fun to work on.

I made comments and changes keeping in mind that the intended audience are Python beginners (and people who don't code). Keep what you want and feel free to discard any part that doesn't feel necessary.

Only two changes in code I made are in the eda notebook.
1. The first plot kept getting cropped on my screen - changed {'bbox_inches': 'tight'}
2. Changed csv file name, and provided link to the csv file

The only part that I think you might want to revisit is the HP-tuning section. Save strategy is set to no but load_best_model_at_end param is explicitly set to True. Also, is there a reason we are not keeping the best performing model but re-finetuning after the HP-tuning process?

I think everything else was clear. Let me know what you think. I'll revisit the notebooks and see if I can add more comments in a few days.